lesson 28 - smart line items

+++ src/routes/(dashboard)/invoices/LineItemRow.svelte
@@ -1,13 +1,20 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
 	import Trash from '$lib/components/icon/Trash.svelte';
-	import { numToCurrency, currencyToNum, twoDecimals } from '$lib/utils/moneyHelpers';
+	import { twoDecimals } from '$lib/utils/moneyHelpers';
 
 	const dispatch = createEventDispatcher();
 
 	export let lineItem: LineItems;
+	export let canDelete: boolean = false;
 
 	let unitPrice: string = twoDecimals(lineItem.amount / lineItem.quantity);
+	let amount: string = twoDecimals(lineItem.amount);
+
+	$: {
+		amount = twoDecimals(lineItem.quantity * Number(unitPrice));
+		lineItem.amount = Number(amount);
+	}
 </script>
 
 <div class="py-2 border-b-2 invoice-line-item border-fog">
@@ -44,17 +51,20 @@
 			name="amount"
 			step="0.01"
 			min="0"
-			bind:value={lineItem.amount}
+			bind:value={amount}
+			disabled
 		/>
 	</div>
 
 	<div>
-		<button
-			on:click|preventDefault={() => {
-				dispatch('removeLineItem', lineItem.id);
-			}}
-			class="w-10 h-10 center text-pastelPurple hover:text-lavenderIndigo"><Trash /></button
-		>
+		{#if canDelete}
+			<button
+				on:click|preventDefault={() => {
+					dispatch('removeLineItem', lineItem.id);
+				}}
+				class="w-10 h-10 center text-pastelPurple hover:text-lavenderIndigo"><Trash /></button
+			>
+		{/if}
 	</div>
 </div>
 
@@ -75,5 +85,10 @@
 	input[type='text']:focus,
 	input[type='number']:focus {
 		@apply border-solid border-lavenderIndigo outline-none;
+	}
+
+	input[type='number']:disabled,
+	input[type='text']:disabled {
+		@apply border-b-0 bg-transparent px-0;
 	}
 </style>
